Back at it after finals
Back to working on this project after midterms. Working on the database connector as of right now. Once that is finished Im getting back to the command processing on the beacon side and the sending from the c2. After all of this its time to work on the panic mode so  that I alwats have a way back in if the binary isnt discovered