Getting some work done
Grabbing a DB connector from friend because I cant be bothered to reinvent the wheel. Going to finish up alot of this tomorrow. Having the whole afternoon to work tomorrow is going to be nice as I can finish this on top of the homework and other stuff I have todo